requirements.txt
flask==3.1.0
line-bot-sdk==3.14.2
google-generativeai==0.8.3
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
gunicorn